requirements.txt
torch
torchvision
pillow
opencv-python
numpy
pandas
scikit-learn
tqdm
matplotlib
seaborn
streamlit
kaggle
git+https://github.com/openai/CLIP.git
ftfy
regex
